feat: add support for `kalnoy/nestedset`
