+++ composer.json
@@ -23,6 +23,7 @@
         "php": "^8.1",
         "filament/filament": "^3.0",
         "illuminate/contracts": "^10.0",
+        "kalnoy/nestedset": "^6.0",
         "spatie/laravel-package-tools": "^1.15.0",
         "staudenmeir/laravel-adjacency-list": "^1.18"
     },
