+++ src/Forms/Components/Actions/AddAction.php
@@ -6,7 +6,7 @@ use Filament\Forms\Components\Actions\Action;
 use Filament\Forms\Form;
 use Filament\Support\Enums\ActionSize;
 use Illuminate\Support\Str;
-use Saade\FilamentAdjacencyList\Forms\Components\AdjacencyList;
+use Saade\FilamentAdjacencyList\Forms\Components\Component;
 
 class AddAction extends Action
 {
@@ -28,7 +28,7 @@ class AddAction extends Action
         $this->modalSubmitActionLabel(fn (): string => __('filament-adjacency-list::adjacency-list.actions.add.modal.actions.create'));
 
         $this->action(
-            function (AdjacencyList $component, array $data): void {
+            function (Component $component, array $data): void {
                 $items = $component->getState();
 
                 $items[(string) Str::uuid()] = [
@@ -44,11 +44,11 @@ class AddAction extends Action
         $this->size(ActionSize::Small);
 
         $this->form(
-            fn (AdjacencyList $component, Form $form) => $component->getForm($form)
+            fn (Component $component, Form $form) => $component->getForm($form)
         );
 
         $this->visible(
-            fn (AdjacencyList $component): bool => $component->isAddable()
+            fn (Component $component): bool => $component->isAddable()
         );
     }
 }
