+++ src/Forms/Components/Actions/EditAction.php
@@ -2,17 +2,13 @@
 
 namespace Saade\FilamentAdjacencyList\Forms\Components\Actions;
 
-use Filament\Actions\Concerns\InteractsWithRecord;
-use Filament\Actions\Contracts\HasRecord;
 use Filament\Forms\Components\Actions\Action;
 use Filament\Forms\Form;
 use Filament\Support\Enums\ActionSize;
-use Saade\FilamentAdjacencyList\Forms\Components\AdjacencyList;
+use Saade\FilamentAdjacencyList\Forms\Components\Component;
 
-class EditAction extends Action implements HasRecord
+class EditAction extends Action
 {
-    use InteractsWithRecord;
-
     public static function getDefaultName(): ?string
     {
         return 'edit';
@@ -31,7 +27,7 @@ class EditAction extends Action implements HasRecord
         $this->modalSubmitActionLabel(fn (): string => __('filament-adjacency-list::adjacency-list.actions.edit.modal.actions.save'));
 
         $this->action(
-            function (AdjacencyList $component, array $arguments, array $data): void {
+            function (Component $component, array $arguments, array $data): void {
                 $statePath = $component->getRelativeStatePath($arguments['statePath']);
                 $state = $component->getState();
 
@@ -46,49 +42,17 @@ class EditAction extends Action implements HasRecord
         $this->size(ActionSize::Small);
 
         $this->form(
-            fn (AdjacencyList $component, Form $form) => $component->getForm($form)
+            fn (Component $component, Form $form) => $component->getForm($form)
         );
 
         $this->fillForm(
-            function (AdjacencyList $component, array $arguments) {
+            function (Component $component, array $arguments) {
                 return data_get($component->getState(), $component->getRelativeStatePath($arguments['statePath']), []);
             }
         );
 
-        $this->record(
-            fn (AdjacencyList $component, array $arguments) => $component->getCachedExistingRecords()->firstWhere($component->getPath(), $component->getRelativeStatePath($arguments['statePath']))
-        );
-
         $this->visible(
-            fn (AdjacencyList $component): bool => $component->isEditable()
+            fn (Component $component): bool => $component->isEditable()
         );
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    protected function resolveDefaultClosureDependencyForEvaluationByName(string $parameterName): array
-    {
-        return match ($parameterName) {
-            'record' => [$this->getRecord()],
-            default => parent::resolveDefaultClosureDependencyForEvaluationByName($parameterName),
-        };
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    protected function resolveDefaultClosureDependencyForEvaluationByType(string $parameterType): array
-    {
-        $record = $this->getRecord();
-
-        if (! $record) {
-            return parent::resolveDefaultClosureDependencyForEvaluationByType($parameterType);
-        }
-
-        return match ($parameterType) {
-            Model::class, $record::class => [$record],
-            default => parent::resolveDefaultClosureDependencyForEvaluationByType($parameterType),
-        };
     }
 }
