+++ src/Forms/Components/Concerns/HasRelationship.php
@@ -1,0 +1,195 @@
+<?php
+
+namespace Saade\FilamentAdjacencyList\Forms\Components\Concerns;
+
+use Closure;
+use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\Relation;
+
+trait HasRelationship
+{
+    protected string | Closure | null $relationship = null;
+
+    protected ?Collection $cachedExistingRecords = null;
+
+    protected string | Closure | null $orderColumn = null;
+
+    protected ?Closure $modifyRelationshipQueryUsing = null;
+
+    protected ?Closure $mutateRelationshipDataBeforeCreateUsing = null;
+
+    protected ?Closure $mutateRelationshipDataBeforeFillUsing = null;
+
+    protected ?Closure $mutateRelationshipDataBeforeSaveUsing = null;
+
+    public function relationship(string | Closure $name = null, Closure $modifyQueryUsing = null): static
+    {
+        $this->relationship = $name ?? $this->getName();
+        $this->modifyRelationshipQueryUsing = $modifyQueryUsing;
+
+        return $this;
+    }
+
+    public function fillFromRelationship(bool $cached = true): void
+    {
+        $this->state(
+            $this->getStateFromRelatedRecords($this->getCachedExistingRecords(cached: $cached)),
+        );
+    }
+
+    /**
+     * @return array<array<string, mixed>>
+     */
+    protected function getStateFromRelatedRecords(Collection $records): array
+    {
+        if (! $records->count()) {
+            return [];
+        }
+
+        return $records
+            ->toTree()
+            ->mapWithKeys(
+                $cb = function (Model $record) use (&$cb): array {
+                    $childrenKey = $this->getChildrenKey();
+                    $recordKeyName = $record->getKeyName();
+
+                    $data = $this->mutateRelationshipDataBeforeFill(
+                        $this->getLivewire()->makeFilamentTranslatableContentDriver() ?
+                            $this->getLivewire()->makeFilamentTranslatableContentDriver()->getRecordAttributesToArray($record) :
+                            $record->attributesToArray()
+                    );
+
+                    $data[$childrenKey] = $record->{$childrenKey}->mapWithKeys($cb)->toArray();
+
+                    return ['record-' . $record->{$recordKeyName} => $data];
+                }
+            )
+            ->toArray();
+    }
+
+    public function orderColumn(string | Closure | null $column = 'sort'): static
+    {
+        $this->orderColumn = $column;
+
+        return $this;
+    }
+
+    public function getOrderColumn(): ?string
+    {
+        return $this->evaluate($this->orderColumn);
+    }
+
+    abstract public function getRelationship(): ?Relation;
+
+    public function getRelationshipName(): ?string
+    {
+        return $this->evaluate($this->relationship);
+    }
+
+    public function getCachedExistingRecords(bool $cached = true): Collection
+    {
+        if ($this->cachedExistingRecords && $cached) {
+            return $this->cachedExistingRecords;
+        }
+
+        $relationship = $this->getRelationship();
+        $relationshipQuery = $relationship->getQuery();
+
+        if ($this->modifyRelationshipQueryUsing) {
+            $relationshipQuery = $this->evaluate($this->modifyRelationshipQueryUsing, [
+                'query' => $relationshipQuery,
+            ]) ?? $relationshipQuery;
+        }
+
+        if ($orderColumn = $this->getOrderColumn()) {
+            $relationshipQuery->orderBy($orderColumn);
+        }
+
+        return $this->cachedExistingRecords = $relationshipQuery->get();
+    }
+
+    public function clearCachedExistingRecords(): void
+    {
+        $this->cachedExistingRecords = null;
+    }
+
+    public function getRelatedModel(): string
+    {
+        return $this->getRelationship()->getModel()::class;
+    }
+
+    public function mutateRelationshipDataBeforeCreateUsing(?Closure $callback): static
+    {
+        $this->mutateRelationshipDataBeforeCreateUsing = $callback;
+
+        return $this;
+    }
+
+    /**
+     * @param  array<array<string, mixed>>  $data
+     * @return array<array<string, mixed>>
+     */
+    public function mutateRelationshipDataBeforeCreate(array $data): array
+    {
+        if ($this->mutateRelationshipDataBeforeCreateUsing instanceof Closure) {
+            $data = $this->evaluate($this->mutateRelationshipDataBeforeCreateUsing, [
+                'data' => $data,
+            ]);
+        }
+
+        return $data;
+    }
+
+    public function mutateRelationshipDataBeforeSaveUsing(?Closure $callback): static
+    {
+        $this->mutateRelationshipDataBeforeSaveUsing = $callback;
+
+        return $this;
+    }
+
+    /**
+     * @param  array<array<string, mixed>>  $data
+     * @return array<array<string, mixed>>
+     */
+    public function mutateRelationshipDataBeforeFill(array $data): array
+    {
+        if ($this->mutateRelationshipDataBeforeFillUsing instanceof Closure) {
+            $data = $this->evaluate($this->mutateRelationshipDataBeforeFillUsing, [
+                'data' => $data,
+            ]);
+        }
+
+        return $data;
+    }
+
+    public function mutateRelationshipDataBeforeFillUsing(?Closure $callback): static
+    {
+        $this->mutateRelationshipDataBeforeFillUsing = $callback;
+
+        return $this;
+    }
+
+    /**
+     * @param  array<array<string, mixed>>  $data
+     * @return array<array<string, mixed>>
+     */
+    public function mutateRelationshipDataBeforeSave(array $data, Model $record): array
+    {
+        if ($this->mutateRelationshipDataBeforeSaveUsing instanceof Closure) {
+            $data = $this->evaluate(
+                $this->mutateRelationshipDataBeforeSaveUsing,
+                namedInjections: [
+                    'data' => $data,
+                    'record' => $record,
+                ],
+                typedInjections: [
+                    Model::class => $record,
+                    $record::class => $record,
+                ],
+            );
+        }
+
+        return $data;
+    }
+}
